requirements.txt
alive_progress==3.2.0
bnlearn==0.10.2
causal_learn==0.1.3.8
JPype1==1.5.1
networkx==3.4
numpy==2.2.1
openai==1.59.7
pandas==2.2.3
pgmpy==0.1.26
tqdm==4.66.5
